refactor(home): move categoryList to widgets
Related to: #16
