+++ src/widgets/CategoryList/ui/CategoryList.module.scss
