+++ src/widgets/Layout/ui/BottomNavigation/BottomNavigation.module.scss
