+++ src/widgets/Layout/ui/BottomNavigation/BottomNavigation.test.tsx
