+++ src/widgets/Layout/ui/BottomNavigation/index.tsx
