+++ src/widgets/Layout/ui/GNB/GNB.module.scss
