+++ src/widgets/Layout/ui/GNB/GNB.test.tsx
