+++ src/widgets/Layout/ui/GNB/index.tsx
